requirements.txt
streamlit
langchain
chromadb
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
python-pptx
python-dotenv
tqdm
requests
pandas
